SPRINT Tool/requirements.txt
accelerate==1.0.1
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
async-timeout==5.0.1
attrs==24.2.0
bitsandbytes==0.42.0
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==44.0.0
datasets==3.1.0
dill==0.3.8
docstring-parser==0.16
eval-type-backport==0.2.0
filelock==3.16.1
flask==3.0.3
frozenlist==1.5.0
fsspec==2024.9.0
huggingface-hub==0.26.5
idna==3.10
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
packaging==24.2
pandas==2.0.3
peft==0.13.2
propcache==0.2.0
psutil==6.1.0
pyarrow==17.0.0
pycparser==2.22
pygments==2.18.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.4.5
scipy==1.10.1
shtab==1.7.1
six==1.17.0
sympy==1.13.3
tokenizers==0.20.3
torch==2.4.1
tqdm==4.67.1
transformers==4.46.3
triton==3.0.0
trl==0.11.4
typeguard==4.4.0
typing-extensions==4.12.2
tyro==0.9.2
tzdata==2024.2
urllib3==2.2.3
werkzeug==3.0.6
xxhash==3.5.0
yarl==1.15.2
zipp==3.20.2
tree-sitter==0.21.3
tree-sitter-python==0.21.0
faiss-cpu==1.8.0
neo4j==5.25.0